# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lokacara API",
        "description": "API documentation for the Lokacara event management platform. Use the Authorize button at the top to set your Sanctum Bearer token for authenticated endpoints.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organizer/events/{event}/certificates/template": {
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Upload a certificate template image",
                "operationId": "0411979ee5d2d471deea488abc8d2b59",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "template"
                                ],
                                "properties": {
                                    "template": {
                                        "description": "Certificate template image (jpeg/png/jpg, max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Template uploaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Template uploaded successfully"
                                        },
                                        "template_path": {
                                            "type": "string",
                                            "example": "temp/abc123.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (not the owner)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organizer/events/{event}/certificates/distribute": {
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Generate and distribute certificates to all attendees",
                "operationId": "21f17b66c5fd1dce505edb70a170d4ad",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "template_path",
                                    "font_family",
                                    "font_color",
                                    "font_size",
                                    "x_pos",
                                    "is_x_center",
                                    "y_pos",
                                    "is_y_center"
                                ],
                                "properties": {
                                    "template_path": {
                                        "type": "string",
                                        "example": "temp/abc123.jpg"
                                    },
                                    "font_family": {
                                        "type": "string",
                                        "example": "Arial"
                                    },
                                    "font_color": {
                                        "type": "string",
                                        "example": "#000000"
                                    },
                                    "font_size": {
                                        "type": "string",
                                        "example": "Medium",
                                        "enum": [
                                            "Small",
                                            "Medium",
                                            "Large"
                                        ]
                                    },
                                    "x_pos": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "is_x_center": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "y_pos": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "is_y_center": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Certificates being distributed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Certificates are being generated and distributed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (not the owner)"
                    },
                    "404": {
                        "description": "Template file not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events/{event}/certificate": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Download your certificate for an event",
                "operationId": "8d6503c747f60ad6cada9e74a15ed235",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate file downloaded"
                    },
                    "404": {
                        "description": "Not registered, certificate not found, or file missing"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token. Get it by logging in via POST /api/auth/login or POST /api/auth/register.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}